+++ Changelog.md
@@ -1,3 +1,4 @@
 # version 1.0.0
 - First version
-- Generate keys for 257 coins
+- Generate keys for 265 coins
+- New coin support: Cosmos (ATOM)